pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rscope"
version = "0.0.1"
readme = "README.md"
description = "RL training visualizer for Mujoco Playground environments + Brax training"
authors = [
    {name = "Jing Yuan Luo", email = "andrew.luo1@outlook.com"},
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

dependencies = [
    "brax",
    "jax",
    "ml_collections",
    "numpy",
    "paramiko",
    "mujoco",
    "watchdog",
]

keywords = ["mujoco", "reinforcement learning", "brax", "visualization"]

[project.optional-dependencies]
test = [
    "absl-py",
    "pytest",
    "pytest-xdist",
]
dev = [
    "rscope[test]",
    "mypy",
    "pyink",
    "pytype",
    "pre-commit",
    "pytest-xdist",
]
all = [
    "rscope[dev]",
]

[project.urls]
Homepage = "https://github.com/Andrew-Luo1/rscope"

[tool.hatch.build.targets.wheel]
packages = ["rscope"]

[tool.pytest.ini_options]
testpaths = [
    "rscope/",
]

[tool.isort]
force_single_line = true
force_sort_within_sections = true
lexicographical = true
single_line_exclusions = ["typing"]
order_by_type = false
group_by_package = true
line_length = 120
use_parentheses = true
multi_line_output = 3
include_trailing_comma = true

[tool.pyink]
line-length = 80
unstable = true
pyink-indentation = 2
pyink-use-majority-quotes = true

[tool.pytype]
inputs = "."
exclude = [
    "**/*_test.py"
]

output = ".pytype"
report_errors = true

[tool.hatch.build]
include = [
    "rscope/**/*",
]